% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bardisus_scrape.R
\name{clean_author_insts}
\alias{clean_author_insts}
\title{BARD-ISUS Author-Institution cleaner.}
\usage{
clean_author_insts(dat)
}
\arguments{
\item{dat}{is the bard-isus data.frame}
}
\value{
data.frame
}
\description{
Scraping the approved proposal/project information from BARD-ISUS results in a very dirty dataset with respect to the "author" and "institution" information. The typical format appears as: "J. Jones U Alabama", but this is hardly a standard easily remedied. This algorithm first takes a table of strings of institutions to match and replace. Then it assumes what is left over is the author name.
}
