% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_xl}
\alias{custom_xl}
\title{Add sheets to Excel Workbook}
\usage{
custom_xl(wb, sheetnm, dat)
}
\arguments{
\item{wb}{is an openxlsx workbook.}

\item{sheetnm}{is the name of the sheet you want to write.}

\item{dat}{is the tibble or data.frame}
}
\value{
an openxlsx workbook
}
\description{
Add sheets to Excel Workbook
}
