% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus.R
\name{list_elements_to_tibble}
\alias{list_elements_to_tibble}
\title{List to tibbles.}
\usage{
list_elements_to_tibble(l)
}
\arguments{
\item{l}{a list containing only nested character vectors.}
}
\value{
a list composed of tibbles.
}
\description{
Convert nested list items into a single list with multiple tables. This is useful when we run \code{rscopus::abstract_retrieval()} and we want to flatten the \code{content} in the returned list.
}
