% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bardisus_scrape.R
\name{replace_inst}
\alias{replace_inst}
\title{Replace partial string matches of institutions in bard-isus scrape.}
\usage{
replace_inst(vect, insts_)
}
\arguments{
\item{vect}{character. Each element of this vector may include multiple author-affiliation information, eg. "J. Jones UC Davis; S. Smith U Hebrew;".}

\item{insts_}{tibble. The table of data created from DATASET.R as "inst_strings". This table has a non-regex string match for most if not all institutions that might appear in \code{vect}.}
}
\value{
character vector of same length as \code{vect}.
}
\description{
This function helps clean the bard-isus list of approved proposals webscrape, where author information is mixed in with institutional affiliation. The core problem is replacing mismatches at scale (~1750 approved proposals). Furthermore, authors-affiliation information is not consistently presented; there are many cases where author information is partial, or the order in which institutions appear with respect to author is incomplete.
}
