% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_helpers.R
\name{score_BARD}
\alias{score_BARD}
\title{Find and score text for "BARD".}
\usage{
score_BARD(vect, pos = TRUE, brk = NULL)
}
\arguments{
\item{vect}{character vector. The text to be searched.}

\item{pos}{logical. Do we search for signs that the string contains references to US-Isr BARD? Or do we search for "negative" signs that the string refers to some other Bard?}

\item{brk}{character vector. A regex to break up text by. For example, there may be two entries in a single string like: "CSIR; India". We may or may not care. Most often, you would use semicolon with an optional whitespace.}
}
\value{
a numeric vector with the same length of \code{vect}.
}
\description{
Using regular expressions, searches for phrases in the \code{dope1::neg_funds} or \code{dope1::pos_funds}, which is internal data created in data-raw/DATASET.R
}
