% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_str}
\alias{split_str}
\title{Split single string}
\usage{
split_str(string)
}
\arguments{
\item{string}{is a character vector, length 1.}
}
\value{
character vector of variable length.
}
\description{
Split single string
}
