% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_helpers.R
\name{vect_detect}
\alias{vect_detect}
\title{Helper function to detect string in a vector.}
\usage{
vect_detect(vect, rgx)
}
\arguments{
\item{vect}{character.}

\item{rgx}{character. Specifically a regular expression.}
}
\description{
This largely serves as a wraparound for str_detect, but designed for character vectors of length > 1.
}
